Refactor user registration logic and update schema
- Replace UserCreate with AuthRegister in registration endpoint
- Move password hashing to auth_service
- Add RegisterResponse model for registration response
- Remove deprecated UserCreate schema
