+++ app/router/auth_router.py
@@ -1,22 +1,18 @@
 import re
 
 from fastapi import APIRouter, Depends, HTTPException
-from passlib.context import CryptContext
 from sqlalchemy import select
 from sqlmodel import Session
 
 from app.database.engine import db_session
 from app.database.models import User
-from app.schema import UserCreate
+from app.schema.auth import AuthRegister, RegisterResponse
+from app.services import auth_service
 
 auth_router = APIRouter(
     prefix="/auth",
     tags=["Authentication"],
 )
-
-def hash_password(password: str) -> str:
-    pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-    return pwd_context.hash(password)
 
 
 def is_valid_email(email: str) -> bool:
@@ -24,8 +20,8 @@ def is_valid_email(email: str) -> bool:
     pattern = r"^[\w\.-]+@[\w\.-]+\.\w+$"
     return re.match(pattern, email) is not None
 
-@auth_router.post("/register")
-def register_user(user_data: UserCreate, session: Session = Depends(db_session)):
+@auth_router.post("/register", response_model=RegisterResponse)
+def register_user(user_data: AuthRegister, session: Session = Depends(db_session)):
     if not is_valid_email(user_data.email):
         raise HTTPException(status_code=400, detail="Invalid email address")
     user_is_registered = session.exec(select(User).where(User.email == user_data.email)).first()
@@ -36,7 +32,7 @@ def register_user(user_data: UserCreate, session: Session = Depends(db_session))
         raise HTTPException(status_code=400, detail="Username already taken")
     if len(user_data.password) < 6:
         raise HTTPException(status_code=400, detail="Password must be at least 6 characters long")
-    hashed_password = hash_password(user_data.password)
+    hashed_password = auth_service.hash_password(user_data.password)
     user = User(
         email=user_data.email,
         password_hash=hashed_password,
@@ -45,4 +41,4 @@ def register_user(user_data: UserCreate, session: Session = Depends(db_session))
     session.add(user)
     session.commit()
     session.refresh(user)
-    return {"message": "Account registered!"}
+    return user
