--- app/schema.py
@@ -1,7 +1,0 @@
-from pydantic import BaseModel
-
-
-class UserCreate(BaseModel):
-    email: str
-    password: str
-    username: str
